Move CLI management service request functions into a separate module
