+++ lib/_cli/provision-config.js
@@ -3,6 +3,7 @@
 var fs = require('fs')
 var ini = require('ini')
 var path = require('path')
+var ManagementService = require('./management-service')
 var util = require('./util')
 var pki = require('./pki')
 
@@ -11,8 +12,8 @@ var PROVISION_COMMAND = 'DxlBrokerMgmt.generateOpenDXLClientProvisioningPackageC
 
 function requestProvisionConfig (hostName, port, userName, password,
                                  csr, verbosity, callback) {
-  util.managementServiceRequest('Requesting client certificate',
-    hostName, port, userName, password, PROVISION_COMMAND,
+  var service = new ManagementService(hostName, port, userName, password)
+  service.request('Requesting client certificate', PROVISION_COMMAND,
     {csrString: csr}, verbosity, callback)
 }
