+++ lib/_cli/update-config.js
@@ -1,17 +1,16 @@
 'use strict'
 
 var path = require('path')
+var ManagementService = require('./management-service')
 var cliUtil = require('./util')
 var dxlUtil = require('../util')
 
 var BROKER_CERT_CHAIN_COMMAND = 'DxlClientMgmt.createClientCaBundle'
 var BROKER_LIST_COMMAND = 'DxlClientMgmt.getBrokerList'
 
-function requestBrokerCertChain (hostName, port, userName, password,
-                                 verbosity, callback) {
-  cliUtil.managementServiceRequest('Requesting broker cert chain',
-    hostName, port, userName, password, BROKER_CERT_CHAIN_COMMAND,
-    {}, verbosity, callback)
+function requestBrokerCertChain (managementService, verbosity, callback) {
+  managementService.request('Requesting broker cert chain',
+    BROKER_CERT_CHAIN_COMMAND, {}, verbosity, callback)
 }
 
 function updateBrokerCertChain (configDir, certChainFile, certChain, verbosity) {
@@ -22,10 +21,8 @@ function updateBrokerCertChain (configDir, certChainFile, certChain, verbosity) 
   cliUtil.saveToFile(caBundleFileName, certChain, verbosity)
 }
 
-function requestBrokerList (hostName, port, userName, password,
-                            verbosity, callback) {
-  cliUtil.managementServiceRequest('Requesting broker list',
-    hostName, port, userName, password, BROKER_LIST_COMMAND,
+function requestBrokerList (managementService, verbosity, callback) {
+  managementService.request('Requesting broker list', BROKER_LIST_COMMAND,
     {}, verbosity, callback)
 }
 
@@ -63,7 +60,7 @@ function updateBrokerListInConfigFile (configFileName, brokerLines, verbosity) {
             // after the Broker section. Comments may pertain to the content of
             // the next section.
             linesAfterBrokerSection.push(line)
-          } else if (line.match(/^\[.*\]$/)) {
+          } else if (line.match(/^\[.*]$/)) {
             // A section after the broker section has been reached.
             inBrokerSection = false
             Array.prototype.push.apply(acc, brokerLines)
@@ -111,18 +108,16 @@ function updateConfig (configDir, hostName, command, verbosity) {
       'BrokerCertChain setting not found in Certs section in config: ' +
       configFileName)
   }
-  requestBrokerCertChain(hostName, command.port, command.username,
-    command.password, verbosity, function (certChain) {
-      updateBrokerCertChain(configDir, config.Certs.BrokerCertChain,
-        certChain, verbosity)
-      requestBrokerList(hostName, command.port, command.username,
-        command.password, verbosity, function (response) {
-          updateBrokerListInConfigFile(configFileName,
-            brokerConfigLines(response), verbosity)
-        }
-      )
-    }
-  )
+  var service = new ManagementService(hostName, command.port,
+    command.username, command.password)
+  requestBrokerCertChain(service, verbosity, function (certChain) {
+    updateBrokerCertChain(configDir, config.Certs.BrokerCertChain,
+      certChain, verbosity)
+    requestBrokerList(service, verbosity, function (response) {
+      updateBrokerListInConfigFile(configFileName,
+        brokerConfigLines(response), verbosity)
+    })
+  })
 }
 
 module.exports = function (program) {
