+++ lib/_cli/util.js
@@ -6,15 +6,10 @@
  */
 
 var fs = require('fs')
-var https = require('https')
 var mkdirp = require('mkdirp')
 var path = require('path')
-var querystring = require('querystring')
 var read = require('read')
-var url = require('url')
 var CliError = require('./cli-error')
-
-var HTTP_MAX_REDIRECTS = 5
 
 var _0644 = parseInt('0644', 8)
 var _0755 = parseInt('0755', 8)
@@ -49,86 +44,6 @@ function throwError (message, component, header) {
   throw new CliError(getErrorMessage(message, component, header))
 }
 
-function buildQueryString (obj) {
-  var query = Object.keys(obj).reduce(function (acc, key) {
-    return acc + '&' + querystring.escape(key) + '=' +
-      querystring.escape(obj[key])
-  }, '')
-  return query.replace(/^&/, '?')
-}
-
-function toUrlString (urlObject) {
-  return 'https://' + urlObject.hostname + ':' + urlObject.port +
-    '/' + urlObject.path
-}
-
-function httpGet (description, currentRedirects, requestOptions, queryString,
-                  verbosity, responseCallback) {
-  var requestUrl = toUrlString(requestOptions)
-  var responseBody = ''
-
-  if (verbosity > 1) {
-    console.log('HTTP request to: ' + requestUrl)
-  }
-
-  var request = https.get(requestOptions, function (response) {
-    if (verbosity > 1) {
-      console.log('HTTP response status code: ' + response.statusCode)
-    }
-    if (response.statusCode === 302) {
-      if (currentRedirects === HTTP_MAX_REDIRECTS) {
-        throwError('Reached maximum redirects for request to ' +
-          requestUrl, description)
-      } else {
-        var redirectLocation = response.headers.location
-        if (!redirectLocation) {
-          throwError('Redirect with no location specified for ' +
-            requestUrl, description)
-        }
-
-        var parsedRedirectLocation = url.parse(redirectLocation)
-        if ((parsedRedirectLocation.hostname &&
-            (parsedRedirectLocation.hostname !== requestOptions.hostname)) ||
-          (parsedRedirectLocation.port &&
-            (parsedRedirectLocation.port !== requestOptions.port))) {
-          throwError('Redirect to different host or port not supported. ' +
-            'Original URL: ' + requestUrl +
-            '. Redirect URL: ' + redirectLocation + '.', description)
-        }
-
-        if (response.headers['set-cookie']) {
-          if (!requestOptions.headers) {
-            requestOptions.headers = {}
-          }
-          requestOptions.headers['cookie'] = response.headers['set-cookie']
-        }
-
-        requestOptions.path = response.headers.location
-        httpGet(description, currentRedirects + 1, requestOptions,
-          queryString, responseCallback, verbosity)
-      }
-      response.resume()
-      return
-    }
-    response.on('data', function (chunk) {
-      responseBody += chunk
-    })
-    response.on('end', function () {
-      if (response.statusCode >= 200 && response.statusCode <= 299) {
-        responseCallback(responseBody)
-      } else {
-        throwError('Request to ' + requestUrl +
-          ' failed with HTTP error code: ' + response.statusCode +
-          '. Reason: ' + response.statusMessage + '.', description)
-      }
-    })
-  })
-  request.on('error', function (error) {
-    throwError('Error processing request to ' + requestUrl + ': ' + error,
-      description)
-  })
-}
-
 module.exports = {
   DEFAULT_CONFIG_FILE_NAME: 'dxlclient.config',
   getProgramVerbosity: function (program) {
@@ -157,49 +72,6 @@ module.exports = {
   saveToFile: function (file, data) {
     module.exports.mkdirRecursive(path.dirname(file))
     fs.writeFileSync(file, data, {mode: _0644})
-  },
-  managementServiceRequest: function (description,
-                                      hostName, port,
-                                      userName, password,
-                                      path, queryString,
-                                      verbosity, responseCallback) {
-    if (typeof queryString === 'undefined') {
-      queryString = {}
-    }
-    queryString[':output'] = 'json'
-
-    var requestOptions = {
-      hostname: hostName,
-      port: port,
-      path: '/remote/' + path + buildQueryString(queryString),
-      rejectUnauthorized: false,
-      requestCert: false
-    }
-    var requestUrl = toUrlString(requestOptions)
-
-    if (userName && password) {
-      requestOptions.auth = userName + ':' + password
-    }
-
-    httpGet(description, 0, requestOptions,
-      queryString, verbosity, function (response) {
-        if (verbosity > 1) {
-          console.log('Response: ' + response)
-        }
-        var responseStatusDelimiter = response.indexOf(':')
-        if (responseStatusDelimiter < 0) {
-          throwError("Did not find ':' status delimiter in response body " +
-            'for request to: ' + requestUrl, description)
-        }
-        var status = response.substring(0, responseStatusDelimiter).trim()
-        var responseDetail = response.substring(
-          responseStatusDelimiter + 1).trim()
-        if (status !== 'OK') {
-          throwError('Request to ' + requestUrl + ' failed with status ' +
-            status + '. Message: ' + responseDetail, description)
-        }
-        responseCallback(JSON.parse(responseDetail))
-      })
   },
   getValueFromPrompt: function (title, confirm, callback) {
     var fullTitle = title.indexOf('Confirm') === 0 ? title : 'Enter ' + title
